End-To-End ML deployment/requirements.txt
mlflow>=2.17,<3.0
boto3>=1.33.3,<2.0
sagemaker>=2.240,<3.0
pandas>=2.2.3,<3.0
numpy>=2.1.0,<3.0
scikit-learn>=1.6.0,<2.0
pyyaml>=6.0.2,<7.0
